Fix pg bug with timespan with time zone and refactor https://github.com/SoftUni-Internal/exam-systems-issues/issues/545
